UAV_DANN/requirements.txt
# UAV-DANN 项目依赖

# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# 配置管理
pyyaml>=6.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.12.0

# 进度条
tqdm>=4.64.0

# TensorBoard (可选)
tensorboard>=2.10.0

# 超参数优化
optuna>=3.0.0

# 其他工具
pillow>=9.0.0
